third_party/blink/web_tests/platform/generic/external/wpt/html/canvas/element/text-styles/2d.text.measure.width.space-expected.txt

This is a testharness.js-based test.
FAIL Space characters are converted to U+0020 and collapsed (per CSS) assert_equals: ctx.measureText('A \x09\x0a\x0c\x0d  \x09\x0a\x0c\x0dB').width === 150 (got 650[number], expected 150[number]) expected 150 but got 650
Harness: the test ran to completion.
